More renaming and stronger spread operator protection
