+++ tests/Rector/MethodCall/AssertDatabaseCountToAssertDatabaseEmptyRector/Fixture/skip_non_test_case_object.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\MethodCall\AssertDatabaseCountZeroToAssertDatabaseEmptyRector\Fixture;
+namespace RectorLaravel\Tests\Rector\MethodCall\AssertDatabaseCountToAssertDatabaseEmptyRector\Fixture;
 
 class SkipNonTestCaseObjectTest
 {
