+++ tests/Rector/MethodCall/AssertDatabaseCountToAssertDatabaseEmptyRector/Fixture/skip_unresolvable_count.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\MethodCall\AssertDatabaseCountZeroToAssertDatabaseEmptyRector\Fixture;
+namespace RectorLaravel\Tests\Rector\MethodCall\AssertDatabaseCountToAssertDatabaseEmptyRector\Fixture;
 
 class SkipUnresolvableCountTest extends \Illuminate\Foundation\Testing\TestCase
 {
