+++ tests/Rector/MethodCall/AssertDatabaseCountToAssertDatabaseEmptyRector/config/configured_rule.php
@@ -3,10 +3,10 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use RectorLaravel\Rector\MethodCall\AssertDatabaseCountZeroToAssertDatabaseEmptyRector;
+use RectorLaravel\Rector\MethodCall\AssertDatabaseCountToAssertDatabaseEmptyRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
 
-    $rectorConfig->rule(AssertDatabaseCountZeroToAssertDatabaseEmptyRector::class);
+    $rectorConfig->rule(AssertDatabaseCountToAssertDatabaseEmptyRector::class);
 };
